+++ symbol_table.c
@@ -1,7 +1,7 @@
 #include "assembler.h"
 
-int16_t search_symbol(char* symbol) {
-    for (int32_t i = 0; i < symbol_table_size; i++) {
+uint64_t search_symbol(char* symbol) {
+    for (uint64_t i = 0; i < symbol_table_size; i++) {
         if (!strcmp(symbol_table[i], symbol)) {
             return i;
         }
